fix: added finding task(s) by owner

+++ task-tracker-api/src/main/java/org/galaxy/tasktrackerapi/model/service/impl/TaskServiceImpl.java
@@ -1,19 +1,20 @@
 package org.galaxy.tasktrackerapi.model.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.galaxy.tasktrackerapi.exception.TaskNotFoundException;
 import org.galaxy.tasktrackerapi.model.dto.TaskCreateDto;
 import org.galaxy.tasktrackerapi.model.dto.TaskReadDto;
 import org.galaxy.tasktrackerapi.model.dto.TaskUpdateDto;
 import org.galaxy.tasktrackerapi.model.entity.Task;
+import org.galaxy.tasktrackerapi.model.entity.User;
 import org.galaxy.tasktrackerapi.model.mapper.TaskMapper;
 import org.galaxy.tasktrackerapi.model.repository.TaskRepository;
 import org.galaxy.tasktrackerapi.model.service.TaskService;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,45 +22,56 @@ import java.util.Optional;
 public class TaskServiceImpl implements TaskService {
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
+    private final MessageSource messageSource;
 
     @Transactional
     @Override
-    public Task createTask(TaskCreateDto taskCreateDto) {
-        return taskRepository.save(taskMapper.taskCreateDtoToTask(taskCreateDto));
+    public TaskReadDto createTask(User user, TaskCreateDto taskCreateDto) {
+        var task = taskMapper.taskCreateDtoToTask(taskCreateDto);
+        task.setUser(user);
+        return taskMapper.taskToTaskReadDto(taskRepository.save(task));
     }
 
     @Override
-    public Optional<Task> findById(Long taskId) {
-        return taskRepository.findById(taskId);
+    public TaskReadDto findByIdAndUser(User user, Long taskId) {
+        return taskMapper.taskToTaskReadDto(taskRepository.findByIdAndUser(taskId, user).orElseThrow(
+                () -> new TaskNotFoundException(messageSource.getMessage("tasks.errors.not_found",
+                        new Object[0], LocaleContextHolder.getLocale())))
+        );
     }
 
     @Override
-    public List<TaskReadDto> findAllTasks(Boolean iscompleted) {
+    public List<TaskReadDto> findAllTasks(User user, Boolean iscompleted) {
         if (iscompleted == null) {
-             var tasks = (List<Task>)  taskRepository.findAll();
-             return tasks.stream().map(taskMapper::taskToTaskReadDto).toList();
+            var tasks = (List<Task>) taskRepository.findByUser(user);
+            return tasks.stream().map(taskMapper::taskToTaskReadDto).toList();
         } else {
-            var tasks = taskRepository.findAllByIscomplited(iscompleted);
+            var tasks = taskRepository.findByIscomplitedAndUser(iscompleted, user);
             return tasks.stream().map(taskMapper::taskToTaskReadDto).toList();
         }
     }
 
     @Transactional
     @Override
-    public void deleteTask(Long taskId) {
-      taskRepository.deleteById(taskId);
+    public void deleteTask(User user, Long taskId) {
+       if (!taskRepository.findByIdAndUser(taskId, user).isPresent()){
+           throw new TaskNotFoundException(messageSource.getMessage("tasks.errors.not_found",
+                   new Object[0], LocaleContextHolder.getLocale()));
+       } else {
+           taskRepository.deleteById(taskId);
+       }
     }
-
 
 
     @Transactional
     @Override
-    public void updateTask(TaskUpdateDto taskUpdateDto) {
-        taskRepository.findById(taskUpdateDto.getId()).ifPresentOrElse(task -> {
-           task.setTitle(taskUpdateDto.getTitle());
-           task.setDescription(taskUpdateDto.getDescription());
-        },()-> {
-            throw new NoSuchElementException();
+    public void updateTask(User user, Long taskId, TaskUpdateDto taskUpdateDto) {
+        taskRepository.findByIdAndUser(taskId, user).ifPresentOrElse(task -> {
+            task.setTitle(taskUpdateDto.getTitle());
+            task.setDescription(taskUpdateDto.getDescription());
+        }, () -> {
+            throw new TaskNotFoundException(messageSource.getMessage("tasks.errors.not_found",
+                    new Object[0], LocaleContextHolder.getLocale()));
         });
     }
 }
